split wasm and native
