+++ demo/src/lib.rs
@@ -1,0 +1,9 @@
+mod app;
+
+pub use app::MyApp;
+
+#[cfg(target_arch = "wasm32")]
+mod web;
+
+#[cfg(target_arch = "wasm32")]
+pub use web::*;
